requirements.txt
python == 24.2
electrum == 4.5.5
pycryptodome ==3.21.0
zksk == 0.2.2
kyber-py == 0.2.2